Reorganize utils and apply redrawTextBoundingBox.
